feat: Refactor import page to get new UI
